Added details about pokemon encounter
Version is for what version the detail applys.
Method is how to encounter the pokemon.
Chance is how likely you will found the pokemon .
Min-Max level are the range of level you will find for the pokemon.
Condition values are conditions that increases the chance  or make possible to find the pokemon in that area.